Add test.app.properties.example
When running mvn package, a lot of errors are preventing a full build. A
lot of the errors are solved by creating a test.app.properties file for
the application factory to build off of.
Additionally, the RubyGeocodeTest, YahooBossTest, and YahooTest have
been commented out because those services are no longer in use. This
Also prevents errors from calling inactive services.