Fix Address toString breaking geocoding requests
In an effort to provide more detailed logging, I had accidentally broke
geocoding in sage. The toString method is used by the TigerGeoCoderDao
to find things in the db. By changing that method, it could no longer
function as intended. The original toString has been brought back, and
a new method containing the desired logging has been created and is used
in the districtRequest and districtController