Added support for a graceful shutdown on startup
If the app.properties is misconfigured (i.e. yahoo is put as a
registered geocode provider when it is not supported any further) the
app will now gracefully shutdown. The exceptions stack strace can be
seen again with some minor code changes, but logging of the actual error
occurs telling the user where to look.