Fix getProcessByIdTest failing on different data
Previoulsy in the JobProcessDaoTest, the getProcessById Test was very
picky about the expected data. This data is different from dev to prod
and therefore was a bad test. This has been rectified by changing what
the test is looking for while still asserting that the dao has / is
doing its job