Attempt to make tests work on more systems
A serious issue with a lot of the tests for SAGE is that they are
dependent on data being in the database first.

In a few tests I have attempted to make the tests work even without
having data in the database.

Additionally, the tiger timeout tests have been removed. Exceptions
should not appear in the log while testing the application