Fix street type appearing twice in api calls
Previoulsy I made a change to the StreetAddress model class which made
the tests work, and had to potential to make tiger more accurate. After
some intensive testing, that change did not affect the accurracy of the
results and caused an error on some addresses returning the street type
twice. That change has been removed and the test updated to reflect the
result it is actually given.