Update Logging for District & Geocode Controllers
As per issue 6117, I am changing when logging happens so that we have
some way of seeing bad data / data before it is processed or while it is
being processed.

An address that is requested for district assignment or geocoding will
now  be printed to the log file before any work is done on them. After
that The requests are logged in the database, and then work begins on
the request. If that does not fail, the result is then logged.