Break apart the AppFactory
One of the central goals of the Sage spring project has been to
completely remove the AppFactory from SAGE. With this commit, a large
part of that goal has been implemented. However, this change is not over
yet. The AppFactory is still referenced through the code, and the Spring
DI is not complete.