Add logging to help debug Job processes
Recently an issue was occuring with the job process where we the geocode
for a lot of addresses were coming back null. However, there was no
logs to help me investigate the problem. I have added rundudant logging
to certain parts of SAGE to help me investigate this issue further
should it arise again