Create new Geocache api call
There is now an api call that given an adress just like a geocode api
call, will cache the address or update the cached address.

This update also makes geocaching far more robust. In the event that we
have already cached an geocode for an address, on determining we have
a dup key, instead of aborting the update, sage will now try to update
the fields for that row in the db.

This should yeild higher consistency when geocaching