Potentially fix Portugeese translation issue
When first traveling to Sage, a user is greeted with a message from
their browser asking if they want to translate the page from Portugeese.
I have specified the locale as english for all jsp pages. I beleive this
will fix this issue as that message has not appeared locally since the
change