Improve speed of generateMetaData
Part of the geocache dao update to existig cached addresses was to
update the zip5. But updating the zip 5 violates a unique contraint. By
removing the update of the zip5, the errors did not occur, which saves
time during the generateMetaData script. Additionally, this improves the
geocaches ability to update more items in cache because the update is
less likely to fail