Complete work on geocaching admin panel google map
The geocaching admin panels google map section is now complete. There is
a google map that appears in the window over NY on page load and resizes
over the address to be geocached. On change the pins are cleared and the
map resizes over all of NY again. When there are multiple pins on the
same point like (google and geocache) the label of the pins readjusts to
show that instead of gargbling the text. All of this will allow and
admin to determine what geocode is best for an address