Fix screen shake and hovering on pins not working
On the index page, when showing any sort of pin like the district
offices, they title / label of the pin would not appear properly and had
the potential to cause screen shaking. By letting google handle the
hovering by their default way, it completely removes the issues we were
having.