Remove BaseJobController
The BaseJobController has been removed from sage spring because it no
longer serves any purpose. Util classes and the conversion of the job
controllers to spring management remove the need for the have a base
controller.

This also allows us to extend another class should we need to later down
the line