Update geocodes in GeocodeTestBase
The geocodes in geocode test base have been updated to reflect what
tiger returns for the addresses as opposed to the coordinates gained
from google. Additionally the similarity constraint has been lessened
because we are really testing that the application works. We are not
looking for performance out of tiger, since Google's results will always
be better