Remove errors from index.jsp
There are no longer errors in index.jsp. I am now able to call the
environment class and retrieve the properties that are necessary for it
to work. The geocoder active section will need to be updated in the
future. I beleive that section of code can be simplified in someway, or
in a common way like a utils class