Revert DistrictShapeCode to prior state
I incorrectly thought that I would need to change all of the columns,
but I only needed to change the one that I was planning on updating.
Additional research also shows that updating our street files with what
tiger / line currently has will not solve the zip code issue we
currently have. We will need a different source of data for this