Make converting Street Address to Address easier
SAGE already had functionality written to convert an Address object to a
Street Address, but was lacking functionality to do so in reverse. In
certain cases being able to do this easily would help tremendously.

The Address class has a new constructor that takes in a street address,
and a helper method will construct the address in the proper way. This
is much easier than in reverse because a Street Address object will
always have its information split out, where as an Address object could
be split out, or it could be just in the addr1 field