First NYS Geocoder implementation
The app example properties has been updated with a few new params
required to make the new Geocoder function properly. Additionally, the
first implementation of it has been written. Testing will need to be
done to confim url's are being generated successfully, and that the
results are being processed correctly

If SAGE were to be built from this commit, this new geocoder could
never be called. Some additional changes will need to be made
to the Application Factory and the app example properties concerning
ranking, and caching.