Handle non returning geocache queries
Part of the cache refresh script calls the Geocache and determines if we
already have the address cached. In the event that we didn't the script
would crash because I was not properly handling the result set position.
This now allows the script to run without crashing but there is still
some issue with USPS not validating address