Fix USPS validated addrs not being returned
Part of the district assignment process determines whether or not it can
send a usps validated address or if it should use the geocoded address.
This is determined by the quality of the geocode. If the geocode quality
is house or higher then we should use the usps address. However the
previous section of code was only allowing house matches through, so
more accurate results were being left in the dust.