Fix critical bugs on NYSGeocacheRefresh Script
There are a few changes in this commit that allow this script to
function properly as intended.
First, I add logic to close all db connections and allow the script to
exit when it finishes the loop of all of the NYS Geo DB's addresses
Second, The offset math has been changed to allow it to function
properly
Thrid, the logging has been reduced so that less time is spent writing
out and more time can be focused on processing.
Fourth, the limit of addresses queried has ben increased from 1000 to
2000, which reduces the total number of sql calls that need to be made
in the lifetime of the script