Fix multiple bugs with NYSRefreshGeocache script
A few changes have been made to the script,
First, the http request connection is only closed at the end of the
loop. This should help save time by not having to constantly remake the
http connection.
Second, the USPS url string will now remove backticks before sending
them to usps
Thrid, if a call to USPS does fail, the script does not quit out. It
will now continue on with the address it pulled from the NYS Geo DB
table.
Fourth, if a call to USPS fails, additional logging will show the
address in question for easy review later on