Refactor DoBatchGeocache to handle Yahoo only
When this script was originally created, it was with the intention that
we would use google to regeocache everything in our geocache. After we
found out about nys geocoder, it was not necessary to regeocache
everything through google. Now we will only need to regeocache about
112,000 or so that are unaccounted for in the conversion of their
coordinates to our geocache