Initial Restructure of UI
Phase 1 of the UI rework is now complete. The code is now structred in a
way that is easy to follow. Finally, there are no longer single files
with thousands of lines of code.

Phase 2 is testing this new structure and fixing all bugs so that it
functions as the current dev and prod branch