Break out js imports into jsp tags
The jsp files used to contain all of the imports necessary for a single
page, which would be a lot of imports considering all of the different
controllers. By moving all of the imports into a jsp tag, the same
imports can be used on multiple jsps. This is excellent code resuse and
is much easier to maintian going forward