Fix Scrolling not working on admin page
The admin page was no longer scrolling due to a fix that was made to
stop screen shaking on the mian sage page. By changing the css on the
admin page, scrolling ca now appear where it should on the admin page, and not appear
where it shouldnt on the main page.