Change geocache retreival sql
I have changed the sql to not be so dependent on a zip5 parameter. By
utilizing the other aspects of the address, like city state, and other
address information we can still get matches without a zip 5.
Additionally, if we do have a zip5, there is way for that information to
be used. This should help address retrieval whether we use usps
validation or not