Update logging for failed geocache response
Sometimes when specifying a specific provider the failed geocache
message would appear. And while a cache hit was not made, it wasn't
because the cache couldnt or did not hit, it was because the cache was
bypassed. The simple logic update will prevent that error message from
misleading someone analyzing the logs