Improve Zip Dist Assign Results
Currently with sage prod if you search for a zip code on the district
assignment page, you will not get back the zip boundary information.

That issue has several parts to it. First I am ensuring that the zip5
field is populated when checking the geocode quality in the
DistrictShapeFile class. Second, I am ensuring that the best formatted
address is being passed to our geocoders. Third, I am ensuring that if
no state is passed in to google, that it assumes NY is the state. That
allows us to get back postal information which allows the rest of the
district assignment to continue