Create script to batch geocache zip codes
This script just geocaches all of the zip codes we have geometry
information for. Google is the only provider that can handle zip codes,
so all of them are geocoded through google and saved into the geo cache.
This should resolve any other issues with district assigning zip codes
on the main index page.