Fix zip code geocode requests not caching
In the geocode controller the initial input was not being parsed into
the best possible version of the address.
In another commit, the geocache update was changed to use the a input
that could potentially be usps validated. But from a zip code inquiry
that would never happen and the bad address would never get cached.
By simply parsing it into a street address and then back to a regular
address we set all of the fields of the address properly which allows
for proper cahcing