Fix GeoCache Updates failing
The previous sql for the geocache was not properly targeting the right
row. Any zip code without a zip4 would all be overwritten to the latest
set of coordinates that came from that sql. This would cause strange
behavior when looking up zip codes.

By fixing this sql query, it will improve the ability to update a row in
the geocache