Remove geocache lockdown for zip codes
Previously I had changed the district controller such that for a zip
only input, it would only be allowed to hit the geocache. This could
cause confusion for users who specify a provider on the front end. This
change has been reverted back to the way the district controller
operated before that change