Fix null town code retrieval and insert
Fixes and issue where a town code map was checked with the city or town
uppercased, but when going to retrieve it from the map, it is not
uppercased. This would return a null value and would mess up the street
files. With this change many more town codes are now populated.