Update Erie parsing
We receieved a new file for Erie county that contained all of the other
districts that we need information for. That file is now in an xlsx
format and it required changes to the parser and how long the conversion
should last. This helps imporve our ability to have the best information
possible in the streetfile database table