Update Schoharie Parser for new Streetfile
The Schoharie parser has been updated to accomodate the new streetfile
that was sent to us. This new file contains more than just the election
district.

Some minor changes were made to the StreetFinderAddress so that I could
verify that some of the fields were set before writing the line to the
file