Normalize Streetfile data in new parsers
When building SAGE on the streetfile branch with the new data the
StreetFileDaoTest would fail. This was due to a few issues, but all of
them were fixed when the data is normalized to all uppercase and
numerics have their suffix's. This did require a change to the
StreetAddressParser. This will have some affect on the geocache, but
overall it leads to more accurate matching of addresses