Improve script versioning
Some of the scripts require all of the dependencies of SAGE to operate.
This means as the sage version changes, the scripts will also need to
point to the properly versioned war file. Rather than change of the
scripts seperately, there is now one variable that needs to be changed
in the config that handles this