Update tons of constructors to use autowired
In the prior builds of sage, so many services and loggers were just
created on the fly or referenced statically. Spring manages dependencies
in a singleton fashion for us, so it is not necessary / impossible to
keep things constructed the same way. With this commit, sage can
compile, but it cant run due to spring failing to build some
dependencies.