Update database connectors / remove config
The scripts are now talking to the proper database and the base dao has
been given the jbdc templates in preparation of converting all of sage
to using them.
The config class has been removed in several places and replaced by the
environmnt. This wont be possible to completely eliminate until all
database interactions have been updated