Multiple controller fixes
The Sage controllers have had a variety of changes. Most of them were
changes to whether a parameter is required or not, and thier default
values. Other changes included fixing the address service to use the
default provider. Updating some paths for the api calls and some code
refactoring to increase code reuse