Fix job processing failures
On the spring branch, Job processing was not working. With some changes
to the executors that are used and how they are created, and removing
the deprecated method calls the job processor works once again!

Downloading the job file after completion currently does not work, and
it is the final issue with job processing on the spring branch