Fix job file download
The job files were unable to download with the error response "file does
not exist". The problem was that the controller was modifying the name
of the file in 2 ways. First the file extension was being cut off by
spring. 2nd, the slash leading to the final directory was being removed.

The Job functionality of sage spring works as it used to now