Move CachingService & remove config cache property
The Caching Service was moved to a more appropriate location for what it
provides, and how it could be potentially abstracted in the future

The idea of a configurable cache was one of the primary goals of the
spring reword in SAGE but after some investigation it really is not
necessary at all. SAGE uses basically next to nothing to cache all of
its data. The app rpoperty was removed from the example app properties