Fix proivder selection query in SqlRegeocacheDao
A query that returns the geoprovider or the method of an address in the
geocache was failing. BY changing how the result set is handled, the
query works as expected.
A test that should not be run on building the app has been added to
ensure that the query works