Boot SAGE w/ many default values if missing value
SAGE when booting with spring will now have a default value for a
environment property or config property instead of just crashing out.

This does not cover all of the app property values. For example,
database connection info, and smtp mailing properties are not set by
default.

This should help in deployments where the app properties is not
configured properly and it allows us to comment out certain properties
that are not actively being used