Create inital Jmeter test script
I am currently working on a way to perfomance test SAGE SPRING as well
as simulate an actual day of the production environment on this branch.

Included in this commit is a script that can convert the tomcat access
logs into a csv file that can be used by Jmeter for the performance
testing. This allows the use of actual production api calls to be used
in a reliable testing format. Depending on the files placed in
/data/geoapi_data the output csv will change. Any unit of time that has
access logs for (day, week, month, etc) can now be replicated for a
performance test