Fix Sage not caching new results
Due to an error in the Geocache, Sage was unable to determine if outside
providers were cacheable when specified from the app props. This is
because it was comparing a class to an object. By Bridging when an
object is needed versus when a class is needed sage builds and runs as
intended