Handle nys geo less than rooftop accuracy better
NYS Geo records with a quality that is less than rooftop level will now
be sent over to NYSGEO's webservice. The webservice is more up to date
in comparison to the snapshots we download from time to time. Some of
the records that we have downloaded have a better quality on the
webservice. There will be some records that have not yet been updated by
NYS GEO