Improve Address 2 Parsing
Address 2 fields were not being parsed properly through the
StreetAddressParser.

When the address fields are populated, we will accept the addr 2 (which
could lead to some bad data in the Geocache. Addr2 is used for a lot of
things)

When the address has all of information in the address1 field, we will
attempt to match the address 2 with the unit prefix. If not then we can
try one last time when setting the street name.

We will also reorder the address after USPS validates it so that the
addr2 information is in the addr2 field and not at the end of the addr1
information.