Fix and Simplify Map Controller Api
The Map Api controller has been simplified down into a few methods as
opposed to one for each district type.

Additionally, all of the district types are now properly registered in
the database which prevents an insert exception when selecting zips in
the district map ui

All sql files are now in flyway format.