Change OpenLocCode Generation and storage
The OpenLocCodes used to be generated in the response view to the
client. However this caused probelms elsewhere in our admin APIs. The
fix for this was to update the geocode model to generate this field.

We still do not store the OpenLocCode in our database because it can
derived for next to nothing in performance from what is already stored