Fix UI not loading rev geocode and USPS lookup
The divs from USPS and Rev geocoding where inside the district map div.
The dist map div would be hidden when the proper tags are selected for
usps or rev geocoding causing the ui elements to never show up

Additionally, a different section of html was being blocked by some
commenting. This has been fixed as well.

Finally, the menuServicefactory will now use a === instead of a ==