Fix BB Req and Controller commenting
The bluebird request was not being created with the same effects as it
is on dev and prod. Further testing is required but the obvious
differences have been fixed.

Additionally, some of the controller methods had the wrong commenting in
terms of the api path. That has been corrected for the respective
methods