Fix Rev Geo Requests not processing properly
Reverse Geocoding was not working properly. It seemed to always return
insufficient address. I found that since the address was always empty we
were adding "NY" to the state field which helps us match zip codes. I
added a method to see if an address is blank or empty. If the address is
blank on submission then we should leave it that way.