Revamp Api Testing script
The prod rediness test script has been completely revamped to better
support changes in input for the tests. It also tests the success
conditions of a request more thoroughly than a 200 HTTP response.

This script has already helped solve a few issues in the api since I
began changing it