District map intersection bug fixes
The district map overlay viewer needed some changes in regards to all
district intersection searches and same type and intersection type
searches.

Both of these situations are better handled.

Some work needs to be done to disable the intersection check box on
initial view of the page and when all districts are selected