Create script to handle NYS Geo DB method
Recently more and more problems are occuring with data in the geocache
that was most recently updated with a method of 'NYS Geo DB'.

The goal of this script is to send Geocoding requests to SAGE that will
update all of the addresses that have this method.

This should drastically increase the quality of the geocache and
hopefully eliminate all errors steming from this update method