Add fallback prevention and some logging
THe regeocache script had an error where is was able to contact
providers other than nysgeo. This was an expensive mistake that has been
corrected. Now the script will only contact nysgeo and if it does not
get a result then it will not be able to update that particular record.

Also a tad bit of logging was added to the controller so we can more
easily tell when it started and ended. Additionally, the uses of
system.out were replaced with the logger.