Allow cmd line access to admin scripts
We now have the ability to access sections of the code managed by spring
through the command line. This allows us to have api and cmd line access
to admin scripts for easier management.

The pom.xml had several packages updated to do this, as well as some of
the spring configuration itself changed to allow this all to work