Restructure scripts for admin and streetfinder
The admin and streetfinder sections for the project had bash files in
the src/main/java section. Those files are now in the bin directory in
the Geoapi project.

The scripts have also been updated to use a specified log file, which is
configured in their respective properties files.

Additionally, a local key from my system was commited before. That key
has been removed from my local system and from the comments in the code