Ensure Update of Senator Cache after GenMetaData
At the end of the GenerateMetaData script, the incorrect cache was being
rebuilt.

Now the proper cache is being rebuilt, and an admin api utility was
created to manually update the cache anytime we want, and to allow the
command line instance to affect the live running version of the cache.