Rework NYC Street File Parsing
NYC street file parsing had issues with null addresses and zip codes not
being populated.

After some analysis and lots of debugging I have reworked it so that
these issues as far as I can tell are resolved.

Additionally, the script that creates the final sql file was fixed as
well. And some wierd cases where the cardinal direction was not being
picked up because the string was uppercase which has been resolved.