Make pathing configurable for street file parsing
Various changes have been made to the street file scripts. This will
allow the folders for the streetfiles to be configured by the users.
Additionally, it makes the folder structure that the user must create is
clear.
Finally, the masterscript will now also generate the sql file after
running the script.