Update Sage to Spring 5.3.20
A ton of different things broke when we udpated to 5.3.20

I have been working on this for a good while now and I think everything
is now in a working state again. This will require heavy testing +
integration testing.

Ehcache is removed. Shiro is working differently compared to our other
apps.