Adjusting USPS Validation results for dist assign
When district assigning an address the addresses seemed to always come
back as false for USPS validation.

I have attempted to correct this by ensuring the address result will
populate the address base object itself with whatever validated field is
passed into it thru the constructor.

As well as ensured it was set in the addresses in the district
controller in the handleBatchDistrictRequest method