Fix null id in passthrough sometimes
When doing batch geocodes, when the geocache itself was used to get the
geocode, the id was never being reset in the returning object. This
resulted in some addresses just not having an id that was originally
sent in.

This is now set after assembling the address from the geocache which
should solve the problem.

Additionally, I added a note about a certain code block in the usps ams
dao which we may be able to delete in the future